refactor: . Move src/Status* to src/Statuses/
