+++ src/Config/Themes/MinimalThemeCollection.ts
@@ -1,4 +1,4 @@
-import type { StatusCollection } from '../../StatusCollection';
+import type { StatusCollection } from '../../Statuses/StatusCollection';
 
 /**
  * Status supported by the Minimal theme. {@link https://github.com/kepano/obsidian-minimal}
