+++ src/Task.ts
@@ -3,13 +3,13 @@ import type { TaskLocation } from './TaskLocation';
 import type { Recurrence } from './Recurrence';
 import { getSettings, getUserSelectedTaskFormat } from './Config/Settings';
 import { GlobalFilter } from './Config/GlobalFilter';
-import { StatusRegistry } from './StatusRegistry';
-import type { Status } from './Status';
+import { StatusRegistry } from './Statuses/StatusRegistry';
+import type { Status } from './Statuses/Status';
 import { Urgency } from './Urgency';
 import { DateFallback } from './DateFallback';
 import { compareByDate } from './lib/DateTools';
 import { TasksDate } from './Scripting/TasksDate';
-import { StatusType } from './StatusConfiguration';
+import { StatusType } from './Statuses/StatusConfiguration';
 import { TasksFile } from './Scripting/TasksFile';
 import { PriorityTools } from './lib/PriorityTools';
 import { logging } from './lib/logging';
