+++ tests/CustomMatchers/CustomMatchersForFilters.ts
@@ -3,8 +3,8 @@ import type { Task } from '../../src/Task';
 import type { FilterOrErrorMessage } from '../../src/Query/Filter/FilterOrErrorMessage';
 import { fromLine } from '../TestHelpers';
 import { TaskBuilder } from '../TestingTools/TaskBuilder';
-import type { StatusConfiguration } from '../../src/StatusConfiguration';
-import { Status } from '../../src/Status';
+import type { StatusConfiguration } from '../../src/Statuses/StatusConfiguration';
+import { Status } from '../../src/Statuses/Status';
 import { SearchInfo } from '../../src/Query/SearchInfo';
 
 /**
