+++ tests/Query/Filter/FunctionField.test.ts
@@ -4,7 +4,7 @@
 import moment from 'moment';
 
 import { FunctionField } from '../../../src/Query/Filter/FunctionField';
-import { Status } from '../../../src/Status';
+import { Status } from '../../../src/Statuses/Status';
 import type { Task } from '../../../src/Task';
 import { Priority } from '../../../src/Task';
 import {
