+++ tests/Sort.test.ts
@@ -14,8 +14,8 @@ import { DueDateField } from '../src/Query/Filter/DueDateField';
 import { PathField } from '../src/Query/Filter/PathField';
 import { SearchInfo } from '../src/Query/SearchInfo';
 import { Sort } from '../src/Query/Sort';
-import { StatusRegistry } from '../src/StatusRegistry';
-import { StatusConfiguration, StatusType } from '../src/StatusConfiguration';
+import { StatusRegistry } from '../src/Statuses/StatusRegistry';
+import { StatusConfiguration, StatusType } from '../src/Statuses/StatusConfiguration';
 import { SampleTasks, fromLine, fromLines, toLines } from './TestHelpers';
 import { TaskBuilder } from './TestingTools/TaskBuilder';
 import { sortBy } from './TestingTools/SortingTestHelpers';
