+++ tests/StatusSettingsReport.test.ts
@@ -1,6 +1,6 @@
 import { StatusSettings } from '../src/Config/StatusSettings';
-import { tabulateStatusSettings } from '../src/StatusSettingsReport';
-import type { StatusCollection } from '../src/StatusCollection';
+import { tabulateStatusSettings } from '../src/Statuses/StatusSettingsReport';
+import type { StatusCollection } from '../src/Statuses/StatusCollection';
 import { verifyWithFileExtension } from './TestingTools/ApprovalTestHelpers';
 import { coreStatusesData, createStatuses } from './TestingTools/StatusesTestHelpers';
