+++ tests/TestHelpers.ts
@@ -1,8 +1,8 @@
 import { Priority, Task } from '../src/Task';
 import { TaskLocation } from '../src/TaskLocation';
 import { PriorityTools } from '../src/lib/PriorityTools';
-import { Status } from '../src/Status';
-import { StatusType } from '../src/StatusConfiguration';
+import { Status } from '../src/Statuses/Status';
+import { StatusType } from '../src/Statuses/StatusConfiguration';
 import { Recurrence } from '../src/Recurrence';
 import { TaskBuilder } from './TestingTools/TaskBuilder';
